+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [0.1.3] - 2023-11-19
+
+- Fixed: RDS Cluster port-forwarding
+- Improve bastion to resources (EKS, RDS) resource matching based on VPC
+
 ## [0.1.2] - 2023-11-15
 
 - Improve AWS rate limit handling by starting AWS SSM port-forwarding session using boto3 client instead of AWS CLI